[#174] Implement a file change description flow
Implemented a file change description flow
