+++ package.json
@@ -25,6 +25,7 @@
     "nodemon": "^2.0.20"
   },
   "_moduleAliases": {
+    "@config": "config.json",
     "@src": "src",
     "@controllers": "src/controllers",
     "@global": "src/global",
