+++ src/interface/CLI/views/create-pr.js
@@ -44,7 +44,7 @@ async function CreatePRsView({ defaultData }) {
 
                                 if (task) {
                                     ev.setValue('task', task, true);
-                                    return task;
+                                    return ev;
                                 } else {
                                     throw new Error.Log({
                                         name: 'TASK-LOADING-ERROR',
@@ -156,9 +156,9 @@ async function CreatePRsView({ defaultData }) {
                 },
                 {
                     id: 'savePullRequest',
-                    defaultData,
                     text: `Alright, we are ready to start the PR. Let's build it (Y/N)?`,
-                    next: `publishPullRequest`,
+                    next: 'addChangesDescription',
+                    defaultData,
                     events: {
                         onTrigger: async (ev, {print}) => {
                             const isReadyToPR = ev.getValue('isReadyToPR');
@@ -183,12 +183,70 @@ async function CreatePRsView({ defaultData }) {
                     }
                 },
                 {
+                    id: 'addChangesDescription',
+                    text: `Let's start to create the file changes description, can we start (Y/N)? `,
+                    next: 'publishPullRequest',
+                    events: {
+                        onAnswer: async (ev, { boolAnswer, print }, answer) => {
+                            const savedPR = ev.getValue('savedPR');
+                            if (savedPR instanceof Error.Log) {
+                                throw savedPR;
+                            }
+
+                            if (!savedPR) {
+                                print(`The pull request don't exist on the internal database!`);
+                                return ev.trigger();
+                            }
+
+                            return new Promise(async (resolve, reject) => {
+                                const PoolForm = require('../PoolForm');
+                                const newPool = new PoolForm({
+                                    events: {
+                                        onEnd: async (ev) => {
+                                            
+                                            return resolve(ev);
+                                        }
+                                    }
+                                }, ev);
+
+                                for (let index = 0; index < savedPR.fileChanges.length; index++) {
+                                    const change = savedPR.fileChanges[index];
+                                    const next = ((index + 1) < savedPR.fileChanges.length) ? String(index + 1) : '';
+                                    const id = String(index);
+    
+                                    newPool.setQuestion({
+                                        id,
+                                        next,
+                                        text: `\n${change.patch}\n\nFilepath: ${change.filename}\nFile blob: ${change.blob_url}\n\nDescription: `,
+                                        events: {
+                                            onAnswer: async (ev, {}, answer) => {
+                                                change.description = answer;
+                                                return ev;
+                                            }
+                                        }
+                                    });
+                                }
+
+                                await newPool.start();
+                            });
+                        }
+                    }
+                },
+                {
                     id: 'publishPullRequest',
                     defaultData,
+                    next: 'finishing-pr',
                     text: `So if everithing is ok with the saved pull request, can we proceed and publish it (Y/N)?`,
                     events: {
                         onTrigger: async (ev, {print, printTemplate}) => {
                             const savedPR = ev.getValue('savedPR');
+
+                            const updated = await savedPR.updateDescription();
+                            if (updated instanceof Error.Log) {
+                                updated.consolePrint();
+                                return ev.trigger();
+                            }
+
                             if (savedPR) {
                                 print('The PR is ready to be published!');
                                 printTemplate(savedPR);
@@ -199,24 +257,38 @@ async function CreatePRsView({ defaultData }) {
                             const savedPR = ev.getValue('savedPR');
 
                             if (boolAnswer(answer)) {
-                                const publish = await savedPR.publishPR();
-                                if (publish instanceof Error.Log) {
-                                    publish.consolePrint();
+                                const published = await savedPR.publishPR();
+                                if (published instanceof Error.Log) {
+                                    published.consolePrint();
                                     return ev.trigger();
                                 }
 
-                                return ev.setValue('publish', publish);
+                                return ev.setValue('published', published);
                             } else {
                                 return ev.trigger();
                             }
                         }
                     }
+                },
+                {
+                    id: 'finishing-pr',
+                    text: `Would you like to close dev-desk (Y/N)? `,
+                    events: {
+                        onAnswer: async (ev) => {
+                            return ev;
+                        }
+                    }
                 }
             ],
             events: {
-                onEnd: (ev, {printTemplate}) => {
-                    const publish = ev.getValue('publish');
-                    printTemplate(publish);
+                onEnd: (ev, {print}) => {
+                    const published = ev.values.published || {};
+
+                    print(
+                        `The pull request was created, and it's available at the link: ${published.html_url || '--Link not available--'}\n\n`,
+                        'SUCCESS'
+                    );
+                    return published;
                 }
             }
         }
