+++ src/resources/en_US/templates/source/default_commit_description.md
@@ -1,2 +1,3 @@
 File changes:
 ##{{fileChanges:array:FileChange}}##
+
