+++ src/resources/en_US/templates/source/default_file_change.md
@@ -1,2 +1,3 @@
-**##{{filename:string}}##:**
-_##{{description:string}}##_
+> **##{{filename:string}}##:**
+> ##{{description:string}}##
+
